+++ srcs/execution/exec_utils.c
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/30 17:50:54 by smagdela          #+#    #+#             */
-/*   Updated: 2022/04/06 11:03:51 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/04/08 14:46:14 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
