+++ srcs/execution/pipeline.c
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 12:47:56 by smagdela          #+#    #+#             */
-/*   Updated: 2022/04/08 13:01:36 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/04/08 17:56:23 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,63 +65,6 @@ void	file_handler(t_data *data)
 	}
 }
 
-/*
-Launch the execution of the whole token_list, and frees it.
-Reproduce the behavior of multiple pipes, adding file redirectors too.
-*/
-bool	executor(char **envp, t_data *data)
-{
-	t_token	*tmp;
-	pid_t	exit_process;
-	int		wstatus;
-	int		nb_process;
-	int		*pids;
-	int		i;
-
-	if (data->token_list == NULL)
-		return (false);
-	if (data->token_list->next == NULL)
-		fork_exec(data->token_list, envp, data);
-	else
-	{
-		file_handler(data);
-		pipe_handler(data);
-		tmp = data->token_list;
-		exit_process = 0;
-		nb_process = count_cmd(data);
-		pids = malloc(sizeof(int) * nb_process);
-		if (pids == NULL)
-			return (false);
-		i = 0;
-		while (tmp && exit_process != -1 && i < nb_process)
-		{
-			if (tmp->type == WORD && tmp->cmd != NULL)
-			{
-				exit_process = fork_exec(tmp, envp, data);
-				pids[i++] = exit_process;
-			}
-			tmp = tmp->next;
-		}
-		i = -1;
-		while (++i < nb_process)
-			waitpid(pids[i], NULL, 0);
-		wstatus = 0;
-		if (exit_process != -1)
-		{
-			waitpid(exit_process, &wstatus, 0);
-			if (WIFEXITED(wstatus))
-				data->status = WEXITSTATUS(wstatus);
-			if (WIFSIGNALED(wstatus))
-				data->status = 128 + WTERMSIG(wstatus);
-		}
-		else
-			data->status = 1;
-	}
-	if (data->token_list)
-		free_toklist(&data->token_list);
-	return (true);
-}
-
 char	*get_filepath(char **filename)
 {
 	char	*pwd;
@@ -138,4 +81,65 @@ char	*get_filepath(char **filename)
 	free(*filename);
 	*filename = NULL;
 	return (filepath);
+}
+
+static void	executor_aux(char **envp, t_data *data, pid_t *exit_process)
+{
+	t_token	*tmp;
+	int		nb_process;
+	int		i;
+
+	nb_process = count_cmd(data);
+	tmp = data->token_list;
+	*exit_process = 0;
+	i = 0;
+	while (tmp && *exit_process != -1 && i < nb_process)
+	{
+		if (tmp->type == WORD && tmp->cmd != NULL)
+		{
+			tmp->pid = fork_exec(tmp, envp, data);
+			*exit_process = tmp->pid;
+		}
+		tmp = tmp->next;
+	}
+	tmp = data->token_list;
+	while (nb_process != 1 && tmp && tmp->pid != *exit_process)
+	{
+		if (tmp->type == WORD && tmp->pid != -1)
+			waitpid(tmp->pid, NULL, 0);
+		tmp = tmp->next;
+	}
+}
+
+/*
+Launch the execution of the whole token_list, and frees it.
+Reproduce the behavior of multiple pipes, adding file redirectors too.
+*/
+void	executor(char **envp, t_data *data)
+{
+	pid_t	exit_process;
+	int		wstatus;
+
+	if (data->token_list == NULL)
+		return ;
+	if (data->token_list->next == NULL)
+		exit_process = fork_exec(data->token_list, envp, data);
+	else
+	{
+		pipe_handler(data);
+		file_handler(data);
+		executor_aux(envp, data, &exit_process);
+	}
+	wstatus = 0;
+	data->status = 1;
+	if (exit_process != -1)
+	{
+		waitpid(exit_process, &wstatus, 0);
+		if (WIFEXITED(wstatus))
+			data->status = WEXITSTATUS(wstatus);
+		if (WIFSIGNALED(wstatus))
+			data->status = 128 + WTERMSIG(wstatus);
+	}
+	if (data->token_list)
+		free_toklist(&data->token_list);
 }
