+++ srcs/parsing/parsing_utils_2.c
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/04 18:46:59 by smagdela          #+#    #+#             */
-/*   Updated: 2022/04/05 16:55:01 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/04/08 18:04:22 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -90,11 +90,11 @@ void	glue_together(t_token **tmp, t_token **token_list)
 	}
 }
 
-void	expand_remaining_envar(t_token *token_list, t_data *env_data)
+void	expand_remaining_envar(t_data *data)
 {
 	t_token	*tmp;
 
-	tmp = token_list;
+	tmp = data->token_list;
 	while (tmp != NULL)
 	{
 		if (tmp->type == VAR)
@@ -102,8 +102,8 @@ void	expand_remaining_envar(t_token *token_list, t_data *env_data)
 			if (ft_strlen(tmp->data) == 1)
 				tmp->type = WORD;
 			else
-				expand(tmp, env_data);
-			glue_together(&tmp, &token_list);
+				expand(tmp, data);
+			glue_together(&tmp, &data->token_list);
 		}
 		tmp = tmp->next;
 	}
